+++ Sources/NativeLogin/LoginRadiusFacebookLogin.h
@@ -22,13 +22,15 @@
  *  Login
  *
  *  @param controller controller is needed for FB native login
- *  @param params     params should have the permissons array and login behaviour
- *  @param handler    serive completion handler
+ *  @param params    params should have the permissons array and login behaviour
+ *  @param socialAppName  should have unique social app name as a provider in case of multiple social apps for the same provider (eg. facebook_<social app name> )
+ *  @param handler    service completion handler
  */
 
 - (void)loginfromViewController:(UIViewController*)controller
-					 parameters:(NSDictionary*)params
-						handler:(LRAPIResponseHandler)handler;
+                     parameters:(NSDictionary*)params
+ withSocialAppName:(NSString * _Nullable)socialAppName
+                        handler:(LRAPIResponseHandler)handler;
 
 /**
  *  Log out the user
@@ -44,8 +46,8 @@
  *  Call this for facebook login to work properly
  */
 - (BOOL)application:(UIApplication *)application
-			openURL:(NSURL *)url
-  sourceApplication:(NSString *)sourceApplication
-		 annotation:(id)annotation;
+            openURL:(NSURL *)url
+  sourceApplication:(NSString * _Nullable)sourceApplication
+         annotation:(id _Nullable)annotation;
 
 @end